UPDATE after review

+++ src/main/java/com/smartparkinglot/backendservice/services/driverservice/DriverServiceImpl.java
@@ -1,6 +1,9 @@
 package com.smartparkinglot.backendservice.services.driverservice;
 
 import com.smartparkinglot.backendservice.domain.Driver;
+import com.smartparkinglot.backendservice.exceptions.driverexceptions.DriverAlreadyExistsException;
+import com.smartparkinglot.backendservice.exceptions.driverexceptions.DriverNotFoundException;
+import com.smartparkinglot.backendservice.exceptions.driverexceptions.WrongCredentialsException;
 import com.smartparkinglot.backendservice.repositories.DriverRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -8,6 +11,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import javax.management.InstanceAlreadyExistsException;
+import javax.validation.constraints.Null;
 import java.net.HttpURLConnection;
 import java.util.List;
 
@@ -18,47 +22,58 @@ public class DriverServiceImpl implements DriverService {
     DriverRepository driverRepository;
 
     @Override
-    public ResponseEntity<List<Driver>> getAllDrivers() {
-        return new ResponseEntity<List<Driver>>(driverRepository.findAll(),HttpStatus.OK);
+    public List<Driver> getAllDrivers() {
+        List<Driver> driverList = driverRepository.findAll();
+        if (driverList == null){
+            throw new NullPointerException();
+        }
+        return driverList;
     }
 
     @Override
-    public ResponseEntity<Driver> getById(Long id) {
-        return new ResponseEntity<Driver>(driverRepository.findById(id).get(),HttpStatus.OK);
+    public Driver getById(Long id) {
+        return driverRepository.findById(id).orElseThrow(
+                () -> new DriverNotFoundException("Driver not found with id: "+id));
     }
 
     @Override
-    public ResponseEntity<Driver> addOrSave(Driver driver) {
+    public Driver addOrSave(Driver driver) {
         Long driver_id = driver.getId();
         String driver_email = driver.getEmail();
 
         if (driver_id == null && driverRepository.findByEmail(driver_email) != null ){
-            return new ResponseEntity(HttpStatus.CONFLICT); // 409
+            throw new DriverAlreadyExistsException("User already exists with given credentials");
         }else{
-            return new ResponseEntity(driverRepository.save(driver),HttpStatus.OK);
+            return driverRepository.save(driver);
         }
     }
 
     @Override
-    public ResponseEntity deleteDriver(Driver driver) {
-        driverRepository.delete(driver);
-        return new ResponseEntity(HttpStatus.OK);
+    public void deleteDriver(Driver driver) {
+        if (driver.getId() == null)
+            throw new DriverNotFoundException("There is no id attached to driver");
+        try{
+            driverRepository.delete(driver);
+        }catch (Exception ex){
+
+        }
     }
 
+
     @Override
-    public ResponseEntity Login(String email, String password) {
+    public Boolean Login(String email, String password) {
         Driver driver; String driverpassword;
         try{
             driver = driverRepository.findByEmail(email);
             driverpassword = driver.getPassword();
         }catch (Exception e){
-            return new ResponseEntity(HttpStatus.NOT_FOUND);
+            throw new DriverNotFoundException("Driver not found with given credentials");
         }
             if ( driverpassword.equals(password)) {
-                return new ResponseEntity(HttpStatus.ACCEPTED);
+                return true;
             }
             else {
-                return new ResponseEntity(HttpStatus.CONFLICT);
+                throw new WrongCredentialsException("Wrong Credentials");
             }
 
     }
